# API/crystal.api.1.0.0-swagger2.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.01",
    "title": "Crystal Server API",
    "description": "This is a simple Crystal server API. You can find out more about Crystal at <a href='http://getcrystal.net'>http://getcrystal.net</a>  For this sample, you can use the api key 'get-crystal' for tests that need authorization (authorization filters)",
    "license": {
      "name": "MIT",
      "url": "http://github.com/gruntjs/grunt/blob/master/LICENSE-MIT"
    }
  },
  "host": "crystal.getcrystal.net",
  "basePath": "/api/01",
  "securityDefinitions": {},
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/agent/{agentId}/roles": {
      "get": {
        "description": "get agent roles by agent ID",
        "tags": [
          "Agent"
        ],
        "operationId": "getAgentRolesByAgentId_",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "x-is-map": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Role"
              }
            }
          },
          "400": {
            "description": "Error 400"
          }
        }
      }
    },
...
    "/role": {
      "post": {
        "description": "create a new role",
        "tags": [
          "Access"
        ],
        "operationId": "createRole_",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": false,
            "x-is-map": false,
            "schema": {
              "$ref": "#/definitions/Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "Error 400"
          }
        }
      }
    },
    "/roles/{agentId}": {
      "get": {
        "description": "get Roles By AgentId",
        "tags": [
          "Access"
        ],
        "operationId": "getRolesByAgentId_",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "x-is-map": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Role"
              }
            }
          }
        }
      }
    },
    "/governance/send/{fromAddress}/{toAddress}/{token}/{denomination}/{amount}": {
      "post": {
        "description": "sends tokens to from an address to an address",
        "tags": [
          "Governance"
        ],
        "operationId": "Create_sendTokens_",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "fromAddress",
            "in": "path",
            "required": true,
            "x-is-map": false,
            "type": "string",
            "description": "the sender address"
          },
          {
            "name": "toAddress",
            "in": "path",
            "required": true,
            "x-is-map": false,
            "type": "string",
            "description": "the recipient address"
          },
          {
            "name": "amount",
            "in": "path",
            "required": true,
            "x-is-map": false,
            "type": "integer",
            "format": "int64",
            "description": "the amount (TODO: settle on the correct data type)"
          },
          {
            "name": "token",
            "in": "path",
            "required": true,
            "x-is-map": false,
            "type": "string",
            "description": "eth, crystal, etc."
          },
          {
            "name": "denomination",
            "in": "path",
            "required": true,
            "x-is-map": false,
            "type": "string",
            "description": "wei, szabo, etc."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
...
    "/group": {
      "post": {
        "description": "create a new Group",
        "tags": [
          "Access"
        ],
        "operationId": "createGroup_",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "x-is-map": false,
            "schema": {
              "$ref": "#/definitions/Group"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/group/{groupId}/{roleId}": {
      "post": {
        "description": "assign a Role to a Group",
        "tags": [
          "Access"
        ],
        "operationId": "Create_assignRoleToGroup_",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "x-is-map": false,
            "type": "string"
          },
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "x-is-map": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
    "/governance/distribute/{token}/{group}/{formula}": {
      "post": {
        "description": "distributes tokens to a group based on the formula",
        "tags": [
          "Governance"
        ],
        "operationId": "Create_distributeTokens_",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "x-is-map": false,
            "type": "string"
          },
          {
            "name": "group",
            "in": "path",
            "required": true,
            "x-is-map": false,
            "type": "string"
          },
          {
            "name": "formula",
            "in": "path",
            "required": true,
            "x-is-map": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Agent": {
      "title": "Agent",
      "description": "Agent currently represents only users (persons that create logins). In the future will also represent code-agents (contracts, bots, etc.)",
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "lastName": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "userStatus": {
          "description": "User Status - active, inactive, banned",
          "type": "integer",
          "format": "int32"
        },
        "userlogin": {
          "type": "string"
        }
      },
      "required": [
        "id"
      ]
    },
...
    "Group": {
      "title": "Group",
      "description": "container for agents and roles. Groups allow the assignment of agents to roles",
      "type": "object",
      "properties": {
        "Id": {
          "description": "hash",
          "type": "string"
        },
        "agents": {
          "description": "list of agents in this group",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Agent"
          }
        },
        "groupName": {
          "type": "string"
        },
        "roles": {
          "description": "list of roles allowed for agents in this group",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Role"
          }
        }
      },
      "required": [
        "Id",
        "agents",
        "groupName"
      ]
    },
    "Role": {
      "title": "Role",
      "description": "a Role represents permissions to a UNC path for CRUD and Execute operations.",
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        },
        "path": {
          "description": "rest endpoint or registered contract endpoint (allows for regex wildcards) to resource(s) that this role can access or execute",
          "type": "string"
        },
        "permissions": {
          "description": "list of permissions for the resources in the path. Agents assigned this role will be allowed to perform GET/POST/DELETE/PATCH (read/update/delete) on the resource.",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Permission"
          }
        }
      },
      "required": [
        "id",
        "name",
        "path",
        "permissions"
      ]
    },
    "Permission": {
      "title": "Permission",
      "description": "a permission for read, write or execute on some unique URI that represents a function or resource. These resources will be the components of the governance scheme, e.g. they will be data or functions that in aggregate comprise a particular governance model.",
      "type": "object",
      "properties": {
        "flags": {
          "description": "create, read, write, execute / or GET/POST",
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        }
      },
      "required": [
        "flags",
        "id",
        "name"
      ]
    },
...
  }
}